# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Church ICT Service Management Platform API",
    "description": "RESTful API for managing ICT support tickets, equipment/room bookings, users, FAQs, announcements, and notifications.",
    "version": "1.0.0"
  },
  "servers": [
    { "url": "/api", "description": "Relative to deployment host" }
  ],
...
  "paths": {
...
    "/tickets": {
      "post": {
        "tags": ["Tickets"],
        "summary": "Submit an ICT support request",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": {
            "type": "object",
            "required": ["category", "description"],
            "properties": {
              "category": { "type": "string", "example": "Projector issue" },
              "description": { "type": "string" },
              "priority": { "type": "string", "enum": ["Low", "Normal", "High", "Urgent"] }
            }
          } } }
        },
        "responses": { "201": { "description": "Ticket created" }, "400": { "description": "Missing fields" } }
      },
      "get": {
        "tags": ["Tickets"],
        "summary": "List tickets (own tickets for Users; all tickets for Staff/Admin)",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "status", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "List of tickets" } }
      }
    },
    "/tickets/{id}": {
      "get": {
        "tags": ["Tickets"],
        "summary": "Get a single ticket",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Ticket" }, "404": { "description": "Not found" } }
      }
    },
    "/tickets/{id}/assign": {
      "patch": {
        "tags": ["Tickets"],
        "summary": "Assign a ticket to ICT staff (Staff/Admin only)",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": {
            "type": "object", "required": ["assigned_to"], "properties": { "assigned_to": { "type": "integer" } }
          } } }
        },
        "responses": { "200": { "description": "Ticket assigned" }, "403": { "description": "Forbidden" } }
      }
    },
    "/tickets/{id}/status": {
      "patch": {
        "tags": ["Tickets"],
        "summary": "Update ticket status through the workflow (Staff/Admin only)",
        "description": "Valid transitions: Pending→Assigned, Assigned→In Progress, In Progress→Resolved, Resolved→Closed (plus limited backward moves).",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": {
            "type": "object", "required": ["status"],
            "properties": { "status": { "type": "string", "enum": ["Pending", "Assigned", "In Progress", "Resolved", "Closed"] } }
          } } }
        },
        "responses": { "200": { "description": "Status updated" }, "400": { "description": "Invalid transition" } }
      }
    },
...
    "/items": {
      "get": {
        "tags": ["Items"],
        "summary": "List equipment and rooms",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "name": "type", "in": "query", "schema": { "type": "string", "enum": ["Equipment", "Room"] } }],
        "responses": { "200": { "description": "List of items" } }
      },
      "post": {
        "tags": ["Items"],
        "summary": "Add new equipment/room (Admin only)",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": {
            "type": "object", "required": ["name", "type"],
            "properties": {
              "name": { "type": "string" },
              "type": { "type": "string", "enum": ["Equipment", "Room"] },
              "description": { "type": "string" }
            }
          } } }
        },
        "responses": { "201": { "description": "Created" } }
      }
    },
    "/items/{id}": {
      "patch": {
        "tags": ["Items"],
        "summary": "Update item details/status (Admin only)",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Updated" } }
      },
      "delete": {
        "tags": ["Items"],
        "summary": "Delete an item (Admin only)",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "204": { "description": "Deleted" } }
      }
    },
...
    "/notifications": {
      "get": {
        "tags": ["Notifications"],
        "summary": "List the logged-in user's notifications",
        "security": [{ "BearerAuth": [] }],
        "responses": { "200": { "description": "List of notifications" } }
      }
    },
    "/notifications/{id}/read": {
      "patch": {
        "tags": ["Notifications"],
        "summary": "Mark a notification as read",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Updated" } }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" }
    }
  }
}